[MIG] auth_brute_force: Migration to 10.0 (OCA/server-tools#877)
