+++ auth_brute_force/__init__.py
@@ -1,3 +1,4 @@
 # -*- encoding: utf-8 -*-
+
 from . import models
 from . import controllers
