+++ auth_brute_force/controllers/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
-from . import controllers
+
+from . import main
