+++ auth_brute_force/data/ir_config_parameter.xml
@@ -1,24 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!-- ********************************************************************** -->
-<!--Tracks Authentication Attempts and Prevents Brute-force Attacks module  -->
-<!--Copyright (C) 2015-Today GRAP (http://www.grap.coop)                    -->
-<!--@author Sylvain LE GAL (https://twitter.com/legalsylvain)               -->
+<!-- Copyright 2015 GRAP -Sylvain LE GAL
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo noupdate="1">
 
-<!--This program is free software: you can redistribute it and/or modify    -->
-<!--it under the terms of the GNU Affero General Public License as          -->
-<!--published by the Free Software Foundation, either version 3 of the      -->
-<!--License, or (at your option) any later version.                         -->
-
-<!--This program is distributed in the hope that it will be useful,         -->
-<!--but WITHOUT ANY WARRANTY; without even the implied warranty of          -->
-<!--MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the            -->
-<!--GNU Affero General Public License for more details.                     -->
-
-<!--You should have received a copy of the GNU Affero General Public License-->
-<!--along with this program. If not, see <http://www.gnu.org/licenses/>.    -->
-<!-- ********************************************************************** -->
-<openerp>
-    <data noupdate="1">
+    <data>
 
         <record id="max_attempt_qty" model="ir.config_parameter">
             <field name="key">auth_brute_force.max_attempt_qty</field>
@@ -26,4 +11,5 @@
         </record>
 
     </data>
-</openerp>
+
+</odoo>
