--- auth_brute_force/i18n/auth_brute_force.pot
@@ -1,150 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* auth_brute_force
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 8.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-09-26 00:39+0000\n"
-"PO-Revision-Date: 2015-09-26 00:39+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: auth_brute_force
-#: code:addons/auth_brute_force/models/res_banned_remote.py:75
-#, python-format
-msgid "%s %s - %s %s (ISP: %s)"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.banned.remote,active:0
-msgid "Active"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,attempt_date:0
-msgid "Attempt Date"
-msgstr ""
-
-#. module: auth_brute_force
-#: model:ir.actions.act_window,name:auth_brute_force.action_res_authentication_attempt
-#: model:ir.ui.menu,name:auth_brute_force.menu_res_authentication_attempt
-msgid "Authentication Attempts"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,result:0
-msgid "Authentication Result"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.banned.remote,ban_date:0
-msgid "Ban Date"
-msgstr ""
-
-#. module: auth_brute_force
-#: code:addons/auth_brute_force/models/res_authentication_attempt.py:34
-#: view:res.authentication.attempt:auth_brute_force.view_res_authentication_attempt_search
-#: selection:res.authentication.attempt,result:0
-#, python-format
-msgid "Banned"
-msgstr ""
-
-#. module: auth_brute_force
-#: model:ir.actions.act_window,name:auth_brute_force.action_res_banned_remote
-#: model:ir.ui.menu,name:auth_brute_force.menu_res_banned_remote
-msgid "Banned Remotes"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,create_uid:0
-#: field:res.banned.remote,create_uid:0
-msgid "Created by"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,create_date:0
-#: field:res.banned.remote,create_date:0
-msgid "Created on"
-msgstr ""
-
-#. module: auth_brute_force
-#: code:addons/auth_brute_force/models/res_authentication_attempt.py:33
-#: view:res.authentication.attempt:auth_brute_force.view_res_authentication_attempt_search
-#: selection:res.authentication.attempt,result:0
-#, python-format
-msgid "Failed"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,id:0
-#: field:res.banned.remote,id:0
-msgid "ID"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,write_uid:0
-#: field:res.banned.remote,write_uid:0
-msgid "Last Updated by"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,write_date:0
-#: field:res.banned.remote,write_date:0
-msgid "Last Updated on"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.banned.remote,name:0
-msgid "Name"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.banned.remote,description:0
-msgid "Remote Description"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,remote:0
-#: field:res.banned.remote,remote:0
-msgid "Remote ID"
-msgstr ""
-
-#. module: auth_brute_force
-#: view:res.authentication.attempt:auth_brute_force.view_res_authentication_attempt_search
-msgid "Successful"
-msgstr ""
-
-#. module: auth_brute_force
-#: code:addons/auth_brute_force/models/res_authentication_attempt.py:32
-#: selection:res.authentication.attempt,result:0
-#, python-format
-msgid "Successfull"
-msgstr ""
-
-#. module: auth_brute_force
-#: field:res.authentication.attempt,login:0
-msgid "Tried Login"
-msgstr ""
-
-#. module: auth_brute_force
-#: help:res.banned.remote,active:0
-msgid "Uncheck this box to unban the remote"
-msgstr ""
-
-#. module: auth_brute_force
-#: code:addons/auth_brute_force/models/res_banned_remote.py:77
-#, python-format
-msgid "Unidentified Call from %s"
-msgstr ""
-
-#. module: auth_brute_force
-#: view:res.authentication.attempt:auth_brute_force.view_res_authentication_attempt_search
-msgid "Without Success"
-msgstr ""
-
