+++ auth_brute_force/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- encoding: utf-8 -*-
+
 from . import res_banned_remote
 from . import res_authentication_attempt
