+++ auth_brute_force/models/res_authentication_attempt.py
@@ -1,27 +1,8 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    Tracks Authentication Attempts and Prevents Brute-force Attacks module
-#    Copyright (C) 2015-Today GRAP (http://www.grap.coop)
-#    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# -*- coding: utf-8 -*-
+# Copyright 2015 GRAP - Sylvain LE GAL
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from openerp import models, fields, api
-from openerp.tools.translate import _
+from odoo import _, api, fields, models
 
 
 class ResAuthenticationAttempt(models.Model):
@@ -36,11 +17,8 @@ class ResAuthenticationAttempt(models.Model):
 
     # Column Section
     attempt_date = fields.Datetime(string='Attempt Date')
-
     login = fields.Char(string='Tried Login')
-
     remote = fields.Char(string='Remote ID')
-
     result = fields.Selection(
         selection=_ATTEMPT_RESULT, string='Authentication Result')
