+++ auth_brute_force/views/action.xml
@@ -1,24 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!-- ********************************************************************** -->
-<!--Tracks Authentication Attempts and Prevents Brute-force Attacks module  -->
-<!--Copyright (C) 2015-Today GRAP (http://www.grap.coop)                    -->
-<!--@author Sylvain LE GAL (https://twitter.com/legalsylvain)               -->
-
-<!--This program is free software: you can redistribute it and/or modify    -->
-<!--it under the terms of the GNU Affero General Public License as          -->
-<!--published by the Free Software Foundation, either version 3 of the      -->
-<!--License, or (at your option) any later version.                         -->
-
-<!--This program is distributed in the hope that it will be useful,         -->
-<!--but WITHOUT ANY WARRANTY; without even the implied warranty of          -->
-<!--MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the           -->
-<!--GNU Affero General Public License for more details.                     -->
-
-<!--You should have received a copy of the GNU Affero General Public License-->
-<!--along with this program.  If not, see <http://www.gnu.org/licenses/>.   -->
-<!-- ********************************************************************** -->
-<openerp>
-    <data>
+<!-- Copyright 2015 GRAP -Sylvain LE GAL
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo>
 
         <record id="action_res_authentication_attempt" model="ir.actions.act_window">
             <field name="name">Authentication Attempts</field>
@@ -35,5 +18,4 @@
             <field name="view_mode">tree,form</field>
         </record>
 
-    </data>
-</openerp>
+</odoo>
